Add custom instructions for GitHub Copilot
- Comprehensive guidelines for Design Patterns repository
- Support for multiple languages (Java, Python, TypeScript)
- Educational focus with Vietnamese comments
- Pattern-specific implementation guidelines
- Consistent coding standards across all patterns
