+++ Creational/Singleton/Java/Example/DatabaseConnection.java
