+++ Creational/Singleton/Java/Structure/EagerSingleton.java
