+++ Creational/Singleton/Php/Example/DatabaseConnection.php
