+++ Creational/Singleton/Php/Structure/EagerSingleton.php
