+++ Creational/Singleton/Php/Structure/LazySingleton.php
